feat(phase1): comprehensive audit fixes across all features
POS / Stock:
- POS.jsx: fix crash — setPaid() -> setPayment() in offline fallback path
- web_pos.py: fix float qty totals (int() was truncating 0.5kg -> 0)

Business type labels:
- biz_language.py: add clinic group (patient lang) + food group (order lang)
- bizLabels.js: add clinic (Patient) + food labels so health businesses
  see Patient not Customer everywhere
- Capture.jsx: SaleForm + PaymentForm now use getBizLabels so Customer
  label adapts per business type (Patient/Client/Member/Student)

Thrift:
- web_routes.py: fix 500 crash on Add Participant — column is customer_phone
- Thrift.jsx: replace PiggyBank icon with Landmark (bank icon)
- Thrift.jsx: fix white text on white card background (ThriftExplainer)
- Thrift.jsx: rename Go to Wallet -> Connect Wallet

Automation:
- index.css: auto-save-bar sticky + right-aligned, clears tiTi FAB on mobile

Security:
- web_routes.py: notes GET — staff see only visibility=all notes (was all)
- web_routes.py: notes DELETE — use resolved owner_phone not session phone

Suppliers:
- supplier_routes.py: ENTERPRISE plan can apply for verified supplier listing

+++ web_pos.py
@@ -16,7 +16,7 @@ def save_pos_sale(db, owner_phone, user_id, customer_id, items, payment_amount, 
 
     Returns receipt dict.
     """
-    total = sum(int(it.get("qty", 1)) * int(it.get("unit_price", 0)) for it in items)
+    total = sum(float(it.get("qty", 1)) * int(it.get("unit_price", 0)) for it in items)
     paid = min(int(payment_amount or 0), total)
 
     has_customer = bool(customer_id)
